Unet + parameters of models
I add a Unet model which goes from 64 to 32 and back to 64 before entering in the neural network. I also add the neural network class to use it in every other model. I added some parameters to tune such as n1, n2, n3  for the numbers of neurons in the neural network. I added one layer in the Neural Network.